offer 30 percent done
offer 30 percent done

+++ admin/offers/offertable.php
@@ -22,18 +22,20 @@
             if($offers_data = $offersObj->fetch()){
                 $counter = 1;
                 foreach ($offers_data as $key => $value) {
-                    echo '<tr>';
-                    echo '<td class="d-lg-none"></td>';
-                    echo '<td scope="row" class="text-center d-none d-sm-table-cell">'; echo_safe($counter); echo'</td>';
-                    echo ' <td>'; echo_safe($value['offer_name']); echo '</td>';
-                    echo '<td class="text-center ">';echo_safe($value['type_of_subscription_details']); echo '</td>';
-                    echo '<td class="text-center ">';echo_safe($value['age_qualification_details']); echo '</td>';
-                    echo '<td class="text-center ">';echo_safe($value['offer_duration']); echo '</td>';
-                    echo '<td class="text-center ">';echo_safe($value['offer_slots']); echo '</td>';
-                    echo '<td class="text-center ">';echo_safe($value['offer_price']); echo '</td>';
-                    echo '<td class="text-center "><a href="editoffer.php?id='; echo_safe($value['offer_id']); echo'" class="btn btn-primary btn-sm" role="button">Edit</a> <a href="deleteoffer.php?id='; echo_safe($value['offer_id']); echo'" class="btn btn-danger btn-sm">Delete</a></td>';
-                    echo '</tr>';
-                    $counter++;
+                    if($value['status_details'] =='active'){
+                        echo '<tr>';
+                        echo '<td class="d-lg-none"></td>';
+                        echo '<td scope="row" class="text-center d-none d-sm-table-cell">'; echo_safe($counter); echo'</td>';
+                        echo ' <td>'; echo_safe($value['offer_name']); echo '</td>';
+                        echo '<td class="text-center ">';echo_safe($value['type_of_subscription_details']); echo '</td>';
+                        echo '<td class="text-center ">';echo_safe($value['age_qualification_details']); echo '</td>';
+                        echo '<td class="text-center ">';echo_safe($value['offer_duration']); echo '</td>';
+                        echo '<td class="text-center ">';echo_safe($value['offer_slots']); echo '</td>';
+                        echo '<td class="text-center ">';echo_safe($value['offer_price']); echo '</td>';
+                        echo '<td class="text-center "><a href="editoffer.php?id='; echo_safe($value['offer_id']); echo'" class="btn btn-primary btn-sm" role="button">Edit</a> <a href="deleteoffer.php?id='; echo_safe($value['offer_id']); echo'" class="btn btn-danger btn-sm">Delete</a></td>';
+                        echo '</tr>';
+                        $counter++;
+                    }
                 }
             }
